% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cnsim.R
\docType{data}
\name{CNSIM3}
\alias{CNSIM3}
\title{Simulated dataset CNSIM 3}
\description{
Simulated dataset CNSIM 1, in a data.frame with 4128 observations of 11 harmonized variables:
LAB_TSC, LAB_TRIG, LAB_HDL, LAB_GLUC_ADJUSTED, PM_BMI_CONTINUOUS, DIS_CVA, MEDI_LPD, DIS_DIAB,
DIS_AMI, GENDER, PM_BMI_CATEGORICAL.
}
\keyword{data}
