% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.dasim.R
\docType{data}
\name{DASIM2}
\alias{DASIM2}
\title{Simulated dataset DASIM 2}
\description{
Simulated dataset DASIM 2, in a data.frame with 10000 observations of 10 harmonized variables:
LAB_TSC, LAB_TRIG, LAB_HDL, LAB_GLUC_FASTING, PM_BMI_CONTINUOUS, DIS_CVA, DIS_DIAB, DIS_AMI,
GENDER, PM_BMI_CATEGORICAL.
}
\keyword{data}
