% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.discordant.R
\docType{data}
\name{DISCORDANT_STUDY2}
\alias{DISCORDANT_STUDY2}
\title{Simulated dataset DISCORDANT 2}
\description{
Simulated dataset DISCORDANT 2, in a data.frame with 12 observations of 2 discordant variables:
A, C.
}
\keyword{data}
