% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsAssignResource,DSLiteConnection-method}
\alias{dsAssignResource,DSLiteConnection-method}
\title{Assign a resource}
\usage{
\S4method{dsAssignResource}{DSLiteConnection}(conn, symbol, resource, async = TRUE)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} object.}

\item{symbol}{Name of the R symbol.}

\item{resource}{Fully qualified name of a resource object living in the DSLite server.}

\item{async}{Whether the result of the call should be retrieved asynchronously. When TRUE (default) the calls are parallelized over
the connections, when the connection supports that feature, with an extra overhead of requests.}
}
\value{
A \code{\linkS4class{DSLiteResult}} object.
}
\description{
Assign a DSLite resource in the DataSHIELD R session.
}
