% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsDisconnect,DSLiteConnection-method}
\alias{dsDisconnect,DSLiteConnection-method}
\title{Disconnect from a DSLite server}
\usage{
\S4method{dsDisconnect}{DSLiteConnection}(conn, save = NULL)
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object}

\item{save}{Save the DataSHIELD R session with provided ID (must be a character string).}
}
\description{
Save the session in a local file if requested.
}
