% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteResult.R
\name{dsGetInfo,DSLiteResult-method}
\alias{dsGetInfo,DSLiteResult-method}
\title{Get result info}
\usage{
\S4method{dsGetInfo}{DSLiteResult}(dsObj, ...)
}
\arguments{
\item{dsObj}{\code{\linkS4class{DSLiteResult}} class object}

\item{...}{Unused, needed for compatibility with generic.}
}
\value{
The result information, including its status.
}
\description{
Get the information about a command (if still available).
}
