% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsHasResource,DSLiteConnection-method}
\alias{dsHasResource,DSLiteConnection-method}
\title{Verify DSLite server resource}
\usage{
\S4method{dsHasResource}{DSLiteConnection}(conn, resource)
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object.}

\item{resource}{The fully qualified name of the resource.}
}
\value{
TRUE if dataset exists.
}
\description{
Verify resource exists and can be accessible for performing DataSHIELD operations.
}
