% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsIsAsync,DSLiteConnection-method}
\alias{dsIsAsync,DSLiteConnection-method}
\title{DSLite asynchronous support}
\usage{
\S4method{dsIsAsync}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} class object}
}
\value{
The named list of logicals detailing the asynchronicity support.
}
\description{
No asynchronicity on any DataSHIELD operations.
}
