% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteResult.R
\name{dsIsCompleted,DSLiteResult-method}
\alias{dsIsCompleted,DSLiteResult-method}
\title{Get whether the operation is completed}
\usage{
\S4method{dsIsCompleted}{DSLiteResult}(res)
}
\arguments{
\item{res}{\code{\link{DSLiteResult-class}} object.}
}
\value{
Always TRUE.
}
\description{
Always TRUE because of synchronous operations.
}
