% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsKeepAlive,DSLiteConnection-method}
\alias{dsKeepAlive,DSLiteConnection-method}
\title{Keep connection with a DSLite server alive}
\usage{
\S4method{dsKeepAlive}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} class object}
}
\description{
No operation due to the DSLiteServer nature.
}
