% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\docType{methods}
\name{dsListMethods,DSLiteConnection-method}
\alias{dsListMethods,DSLiteConnection-method}
\title{List methods}
\usage{
\S4method{dsListMethods}{DSLiteConnection}(conn, type = "aggregate")
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\value{
A data frame.
}
\description{
List methods defined in the DataSHIELD configuration.
}
