% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsListPackages,DSLiteConnection-method}
\alias{dsListPackages,DSLiteConnection-method}
\title{List packages}
\usage{
\S4method{dsListPackages}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} class object}
}
\value{
A data frame.
}
\description{
List packages defined in the DataSHIELD configuration.
}
