% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsListProfiles,DSLiteConnection-method}
\alias{dsListProfiles,DSLiteConnection-method}
\title{List profiles}
\usage{
\S4method{dsListProfiles}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} class object}
}
\value{
A list containing the "available" character vector of profile names and the "current" profile (in case a default one was assigned).
}
\description{
List profiles defined in the DataSHIELD configuration.
}
