% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsListSymbols,DSLiteConnection-method}
\alias{dsListSymbols,DSLiteConnection-method}
\title{List R symbols}
\usage{
\S4method{dsListSymbols}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object}
}
\value{
A character vector.
}
\description{
List symbols living in the DataSHIELD R session.
}
