% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsListTables,DSLiteConnection-method}
\alias{dsListTables,DSLiteConnection-method}
\title{List DSLite server datasets}
\usage{
\S4method{dsListTables}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object}
}
\value{
The fully qualified names of the tables.
}
\description{
List dataset names living in the DSLite server for performing DataSHIELD operations.
}
