% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsListWorkspaces,DSLiteConnection-method}
\alias{dsListWorkspaces,DSLiteConnection-method}
\title{List workspaces}
\usage{
\S4method{dsListWorkspaces}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} class object}
}
\value{
A data frame.
}
\description{
List workspaces saved in the data repository.
}
