% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLite-package.R
\docType{package}
\name{DSLite-package}
\alias{DSLite}
\alias{DSLite-package}
\title{DSLite: 'DataSHIELD' Implementation on Local Datasets}
\description{
'DataSHIELD' is an infrastructure and series of R packages that 
    enables the remote and 'non-disclosive' analysis of sensitive research data.
    This 'DataSHIELD Interface' implementation is for analyzing datasets living
    in the current R session. The purpose of this is primarily for lightweight
    'DataSHIELD' analysis package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datashield/DSLite/}
  \item \url{https://datashield.github.com/DSLite/}
  \item \url{http://www.datashield.org}
  \item \url{https://doi.org/10.1093/ije/dyu188}
  \item Report bugs at \url{https://github.com/datashield/DSLite/issues/}
}

}
\author{
\strong{Maintainer}: Yannick Marcon \email{yannick.marcon@obiba.org} (\href{https://orcid.org/0000-0003-0138-2023}{ORCID})

}
\keyword{internal}
