% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\docType{class}
\name{DSLiteConnection-class}
\alias{DSLiteConnection-class}
\title{Class DSLiteConnection.}
\description{
A DSLiteServer connection implementing the DataSHIELD Interface (DSI) \code{\link[DSI]{DSConnection-class}}.
}
\keyword{internal}
