% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteDriver.R
\docType{class}
\name{DSLiteDriver-class}
\alias{DSLiteDriver-class}
\title{Class DSLiteDriver with constructor DSLite.}
\description{
An DSLite driver implementing the DataSHIELD Interface (DSI) \code{\link[DSI]{DSDriver-class}}.
This class should always be initialized with the \code{\link{DSLite}} function.
It returns a singleton that allows you to connect to DSlite.
}
\keyword{internal}
