% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteResult.R
\docType{class}
\name{DSLiteResult-class}
\alias{DSLiteResult-class}
\title{Class DSLiteResult.}
\description{
An DSLite result implementing the DataSHIELD Interface (DSI) \code{\link[DSI]{DSResult-class}}.
}
\keyword{internal}
