% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsParse.R
\name{dsParse}
\alias{dsParse}
\title{Parse an expression according DataSHIELD syntax rules.}
\usage{
dsParse(expr, debug = rly::NullLogger$new())
}
\arguments{
\item{expr}{Expression}

\item{debug}{Parser debug logger}
}
\value{
An Abstract Syntaxic Tree node
}
\description{
Parse an expression according DataSHIELD syntax rules.
}
