% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsRestoreWorkspace,DSLiteConnection-method}
\alias{dsRestoreWorkspace,DSLiteConnection-method}
\title{Restore workspace}
\usage{
\S4method{dsRestoreWorkspace}{DSLiteConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Restore workspace from the data repository.
}
