% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsRmSymbol,DSLiteConnection-method}
\alias{dsRmSymbol,DSLiteConnection-method}
\title{Remove a R symbol}
\usage{
\S4method{dsRmSymbol}{DSLiteConnection}(conn, symbol)
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object}

\item{symbol}{Name of the R symbol.}
}
\description{
Remoe a symbol living in the DataSHIELD R session.
}
