% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsSaveWorkspace,DSLiteConnection-method}
\alias{dsSaveWorkspace,DSLiteConnection-method}
\title{Save workspace}
\usage{
\S4method{dsSaveWorkspace}{DSLiteConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} class object}

\item{name}{Name of the workspace.}
}
\description{
Save workspace on the data repository.
}
