% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\docType{class}
\name{GroupNode}
\alias{GroupNode}
\title{Group AST node}
\description{
AST node that reprsents a group of tokens enclosed by parenthesis.
}
\seealso{
Other parser items: 
\code{\link{BinaryOpNode}},
\code{\link{FormulaNode}},
\code{\link{FunctionNode}},
\code{\link{Node}},
\code{\link{NumericNode}},
\code{\link{ParameterNode}},
\code{\link{RangeNode}},
\code{\link{StringNode}},
\code{\link{SymbolNode}},
\code{\link{UnaryOpNode}}
}
\concept{parser items}
\section{Super class}{
\code{\link[DSLite:Node]{DSLite::Node}} -> \code{GroupNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupNode-to_string}{\code{GroupNode$to_string()}}
\item \href{#method-GroupNode-clone}{\code{GroupNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="accept"><a href='../../DSLite/html/Node.html#method-Node-accept'><code>DSLite::Node$accept()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="add_child"><a href='../../DSLite/html/Node.html#method-Node-add_child'><code>DSLite::Node$add_child()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="initialize"><a href='../../DSLite/html/Node.html#method-Node-initialize'><code>DSLite::Node$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="set_parent"><a href='../../DSLite/html/Node.html#method-Node-set_parent'><code>DSLite::Node$set_parent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="to_string_children"><a href='../../DSLite/html/Node.html#method-Node-to_string_children'><code>DSLite::Node$to_string_children()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupNode-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-GroupNode-to_string}{}}}
\subsection{Method \code{to_string()}}{
Get the string representation of the GroupNode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupNode$to_string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
