% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cnsim.R
\docType{data}
\name{logindata.dslite.cnsim}
\alias{logindata.dslite.cnsim}
\title{DataSHIELD login data for the CNSIM simulated datasets}
\description{
DataSHIELD login data.frame for connecting with CNSIM datasets.
}
\keyword{data}
