% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.dasim.R
\docType{data}
\name{logindata.dslite.dasim}
\alias{logindata.dslite.dasim}
\title{DataSHIELD login data for the DASIM simulated datasets}
\description{
DataSHIELD login data.frame for connecting with DASIM datasets.
}
\keyword{data}
