% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.discordant.R
\docType{data}
\name{logindata.dslite.discordant}
\alias{logindata.dslite.discordant}
\title{Login data for the DISCORDANT simulated datasets}
\description{
Login data for the DISCORDANT simulated datasets
}
\keyword{data}
