% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.testing.dataset.R
\docType{data}
\name{logindata.dslite.testing.dataset}
\alias{logindata.dslite.testing.dataset}
\title{DataSHIELD login data for the TESTING.DATASET simulated datasets}
\description{
DataSHIELD login data.frame for connecting with TESTING.DATASET datasets which purpose is to
evaluate each base data types.
}
\keyword{data}
