% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDSLiteServer.R
\name{setupDSLiteServer}
\alias{setupDSLiteServer}
\title{Setup an environment based on named datasets and logindata}
\usage{
setupDSLiteServer(packages, datasets, logindata, pkgs = NULL,
  dslite.server = "dslite.server", env = globalenv())
}
\arguments{
\item{packages}{DataSHIELD server-side packages that must be installed locally so that the \link{DSLiteServer}
can execute the DataSHIELD operations.}

\item{datasets}{Names of the datasets to be loaded using \link{data}.}

\item{logindata}{Name of the login data object to be loaded using \link{data}.}

\item{pkgs}{The package(s) to look in for datasets, default is all.}

\item{dslite.server}{Symbol name to which the \link{DSLiteServer} should be assigned to.}

\item{env}{The environment where DataSHIELD objects should be looked for: the \link{DSLiteServer} and
the DSIConnection objects. Default is the Global environment.}
}
\value{
The login data for the \link{datashield.login} function.
}
\description{
Load the provided datasets and the corresponding logindata object, instanciate a new \link{DSLiteServer}
hosting these datasets, verifies that the required DataSHIELD server-side packages are installed.
All the data structures are loaded by \link{data} which supports various formats (see data() documentation).
}
\seealso{
Other setup functions: \code{\link{setupCNSIMTest}}
}
\concept{setup functions}
