% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\docType{class}
\name{SymbolNode}
\alias{SymbolNode}
\title{Symbol AST node}
\description{
AST node that represents a R symbol (variable name, function name etc.).
}
\seealso{
Other parser items: 
\code{\link{BinaryOpNode}},
\code{\link{FormulaNode}},
\code{\link{FunctionNode}},
\code{\link{GroupNode}},
\code{\link{Node}},
\code{\link{NumericNode}},
\code{\link{ParameterNode}},
\code{\link{RangeNode}},
\code{\link{StringNode}},
\code{\link{UnaryOpNode}}
}
\concept{parser items}
\section{Super class}{
\code{\link[DSLite:Node]{DSLite::Node}} -> \code{SymbolNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SymbolNode-add_child}{\code{SymbolNode$add_child()}}
\item \href{#method-SymbolNode-to_string}{\code{SymbolNode$to_string()}}
\item \href{#method-SymbolNode-clone}{\code{SymbolNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="accept"><a href='../../DSLite/html/Node.html#method-Node-accept'><code>DSLite::Node$accept()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="initialize"><a href='../../DSLite/html/Node.html#method-Node-initialize'><code>DSLite::Node$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="set_parent"><a href='../../DSLite/html/Node.html#method-Node-set_parent'><code>DSLite::Node$set_parent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DSLite" data-topic="Node" data-id="to_string_children"><a href='../../DSLite/html/Node.html#method-Node-to_string_children'><code>DSLite::Node$to_string_children()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolNode-add_child"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolNode-add_child}{}}}
\subsection{Method \code{add_child()}}{
No children
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolNode$add_child(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Child Node}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolNode-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolNode-to_string}{}}}
\subsection{Method \code{to_string()}}{
Get the string representation of the SymbolNode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolNode$to_string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
