% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.testing.dataset.R
\docType{data}
\name{TESTING.DATASET1}
\alias{TESTING.DATASET1}
\title{Simulated dataset TESTING.DATASET 1}
\description{
Simulated dataset TESTING.DATASET 1, in a data.frame with 71 observations of 17 harmonized variables:
ID, CHARACTER, LOGICAL, NA_VALUES, NULL_VALUES, INTEGER, NON_NEGATIVE_INTEGER, POSITIVE_INTEGER,
NEGATIVE_INTEGER, NUMERIC, NON_NEGATIVE_NUMERIC, POSITIVE_NUMERIC, NEGATIVE_NUMERIC, FACTOR_CHARACTER,
FACTOR_INTEGER, IDENTIFIER, CATEGORY.
}
\keyword{data}
