% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testParse.R
\name{testParse}
\alias{testParse}
\title{Parse an expression according DataSHIELD syntax rules and returns an Abstract Syntaxic Tree (AST) node.}
\usage{
testParse(expr, debug = FALSE)
}
\arguments{
\item{expr}{Expression}

\item{debug}{Parser debug logger activated}
}
\value{
An Abstract Syntaxic Tree node
}
\description{
Parse an expression according DataSHIELD syntax rules and returns an Abstract Syntaxic Tree (AST) node.
}
