% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteResult.R
\name{dsFetch,DSLiteResult-method}
\alias{dsFetch,DSLiteResult-method}
\title{Fetch the result}
\usage{
\S4method{dsFetch}{DSLiteResult}(res)
}
\arguments{
\item{res}{\code{\link{DSLiteResult-class}} object.}
}
\value{
TRUE if table exists.
}
\description{
Fetch the DataSHIELD operation result.
}
