% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\docType{methods}
\name{dsListResources,DSLiteConnection-method}
\alias{dsListResources,DSLiteConnection-method}
\title{List DSLite server resources}
\usage{
\S4method{dsListResources}{DSLiteConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{DSLiteConnection-class}} class object}
}
\value{
The fully qualified names of the resources.
}
\description{
List resource names living in the DSLite server for performing DataSHIELD operations.
}
