% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteServer.R
\docType{class}
\name{DSLiteServer}
\alias{DSLiteServer}
\title{Lightweight DataSHIELD server-side component}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DSLiteServer
}
\description{
DSLiteServer mimics a DataSHIELD server by holding datasets and exposing
DataSHIELD-like functions: aggregate and assign. A DataSHIELD session is a R
environment where the assignment and the operations happen.
}
\section{Fields}{

\describe{
\item{\code{tables}}{A named list of data.frames representing the harmonized tables.}

\item{\code{config}}{The DataSHIELD configuration. Default is to discover it from the DataSHIELD server-side R packages.}

\item{\code{strict}}{Logical to specify whether the DataSHIELD configuration must be strictly applied. Default is TRUE.}

\item{\code{workspacesDir}}{Folder location where to read and dump workspace images. Default is in a hidden folder of the user home.}
}}

\seealso{
Other server-side items: \code{\link{newDSLiteServer}}
}
\concept{server-side items}
\keyword{datasets}
