% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteConnection.R
\name{dsRmWorkspace,DSLiteConnection-method}
\alias{dsRmWorkspace,DSLiteConnection-method}
\title{Remove a workspace}
\usage{
\S4method{dsRmWorkspace}{DSLiteConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\linkS4class{DSLiteConnection}} class object}

\item{name}{Name of the workspace.}
}
\description{
Remove a workspace on the data repository.
}
