% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.dasim.R
\docType{data}
\name{logindata.dslite.dasim}
\alias{logindata.dslite.dasim}
\title{DataSHIELD login data for the DASIM simulated datasets}
\description{
DataSHIELD login data.frame for connecting with DASIM datasets. The DASIM datasets contain
synthetic data based on a model derived from the participants of the 1958 Birth Cohort, as
part of the obesity methodological development project. These datasets do not contain some
NA values.
}
\details{
\tabular{llll}{
   \strong{Field} \tab \strong{Description} \tab \strong{Type} \tab \strong{Note} \cr
   server \tab Server/study name \tab char \tab  \cr
   url \tab Server/study URL \tab char \tab DSLiteServer instance symbol name \cr
   user \tab User name \tab char \tab Always empty for DSLiteServer \cr
   password \tab User password \tab char \tab Always empty for DSLiteServer \cr
   table \tab Table unique name \tab char \tab As registered in the DSLiteServer \cr
   options \tab Connection options \tab char \tab Always empty for DSLiteServer \cr
   driver \tab Connection driver \tab char \tab DSLiteServer \cr
}
}
\keyword{data}
