% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.discordant.R
\docType{data}
\name{logindata.dslite.discordant}
\alias{logindata.dslite.discordant}
\title{DataSHIELD login data for the DISCORDANT simulated datasets}
\description{
DataSHIELD login data.frame for connecting with DISCORDANT datasets which purpose is to test
datasets that are NOT harmonized.
}
\keyword{data}
