% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.survival.R
\docType{data}
\name{logindata.dslite.survival.expand_with_missing}
\alias{logindata.dslite.survival.expand_with_missing}
\title{Login data for the simulated survival expand-with-missing datasets}
\description{
Login data for the simulated survival expand-with-missing datasets
}
\keyword{data}
