% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.survival.R
\docType{data}
\name{logindata.dslite.survival.expand_with_missing}
\alias{logindata.dslite.survival.expand_with_missing}
\title{DataSHIELD login data for the simulated survival expand-with-missing datasets}
\description{
DataSHIELD login data.frame for connecting with SURVIVAL datasets which purpose is to perform survival tests.
The datasets contain synthetic data based on a simulated survival model, including a censoring indicator.
}
\details{
\tabular{llll}{
   \strong{Field} \tab \strong{Description} \tab \strong{Type} \tab \strong{Note} \cr
   server \tab Server/study name \tab char \tab  \cr
   url \tab Server/study URL \tab char \tab DSLiteServer instance symbol name \cr
   user \tab User name \tab char \tab Always empty for DSLiteServer \cr
   password \tab User password \tab char \tab Always empty for DSLiteServer \cr
   table \tab Table unique name \tab char \tab As registered in the DSLiteServer \cr
   options \tab Connection options \tab char \tab Always empty for DSLiteServer \cr
   driver \tab Connection driver \tab char \tab DSLiteServer \cr
}
}
\keyword{data}
