% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.testing.dataset.R
\docType{data}
\name{logindata.dslite.testing.dataset}
\alias{logindata.dslite.testing.dataset}
\title{Login data for the TESTING.DATASET simulated datasets}
\description{
Login data for the TESTING.DATASET simulated datasets
}
\keyword{data}
