% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLiteServer.R
\name{newDSLiteServer}
\alias{newDSLiteServer}
\title{Create a new DSLite server}
\usage{
newDSLiteServer(tables = list(),
  config = DSLite::defaultDSConfiguration(), strict = TRUE,
  workspacesDir = file.path("~", ".dslite", "workspaces"))
}
\arguments{
\item{tables}{A named list of data.frames representing the harmonized tables.}

\item{config}{The DataSHIELD configuration. Default is to discover it from the DataSHIELD server-side R packages.}

\item{strict}{Logical to specify whether the DataSHIELD configuration must be strictly applied. Default is TRUE.}

\item{workspacesDir}{Folder location where to read and dump workspace images. Default is in a hidden folder of the user home.}
}
\description{
Shortcut function to create a new \link{DSLiteServer} instance.
}
\seealso{
Other server-side items: \code{\link{DSLiteServer}}
}
\concept{server-side items}
