% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.survival.R
\docType{data}
\name{SURVIVAL.EXPAND_WITH_MISSING1}
\alias{SURVIVAL.EXPAND_WITH_MISSING1}
\title{Simulated survival expand-with-missing dataset 1}
\description{
Simulated survival expand-with-missing dataset 1
}
\keyword{data}
