% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.survival.R
\docType{data}
\name{SURVIVAL.EXPAND_WITH_MISSING2}
\alias{SURVIVAL.EXPAND_WITH_MISSING2}
\title{Simulated survival expand-with-missing dataset 2}
\description{
Simulated survival expand-with-missing dataset 2
}
\keyword{data}
