% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.survival.R
\docType{data}
\name{SURVIVAL.EXPAND_WITH_MISSING3}
\alias{SURVIVAL.EXPAND_WITH_MISSING3}
\title{Simulated survival expand-with-missing dataset 3}
\description{
Simulated survival expand-with-missing dataset 3
}
\keyword{data}
